+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## [v1.72.0]
+- Fixed deduplication for dependency-provider rules so dependent validations run per blob
+- Updated Artifactory rule entropy and added new artifactory rule
+
 ## [v1.71.0]
 - Improved Report Viewer layout
 - Improved Salesforce rule